inventory: handle non-weapon mounted image when cycling weapons
If a client attempts to cycle their weapon slot with a non-item image
(e.g. a repair pack is an item; a repair gun is a non-item), this will
generate the following warning on the server and the weapon cycle event
will stop:

    scripts/inventory.cs (...): Unable to find object: '' attempting to call function 'getName'

This commit addresses this by explicitly checking that the image mounted
in the weapon slot references an item type before attempting to find the
current slot on the player.

Signed-off-by: James Knight <james.d.knight@live.com>