commandermap: controller-named turret tag management
An original tweak to the implementation of controller-named turrets
would attempt to avoid the destruction of the original tag to prevent
the need to recreate the tag when a controller relinquished control of
the turret. This change made an assumption the turret entity's "nameTag"
value would always be set (which is not always the case). Adjusting the
implementation to allow a fallback to pull off the tag value from the
turret datablock's "targetNameTag" value (as which was done in classic).

This change also ensures a graceful cleanup of tags by leaving the stock
tag management as it, and only creating a new "controller" tag (i.e. an
interim tag) until the controller releases control or when the entity is
removed.

[1]: 8093dd610ebfbba8b97bb083b7ee3757a460459f

Signed-off-by: James Knight <james.d.knight@live.com>